requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=9.0.0
easyocr>=1.6.2
pandas>=1.3.0
numpy>=1.21.0
pytesseract>=0.3.10